+++ src/culverin/culverin_contact_listener.c
@@ -1,7 +1,6 @@
 #include "culverin_contact_listener.h"
-#include "culverin.h"
 #include "culverin_fast_build.h"
-#include "culverin_physics_sync.h"
+#include "culverin_physics_world.h"
 
 // --- Internal Contact Helper ---
 // NOLINTNEXTLINE(readability-function-cognitive-complexity)
@@ -213,214 +212,3 @@ const JPH_ContactListener_Procs contact_procs = {.OnContactValidate  = on_contac
                                                  .OnContactAdded     = on_contact_added,
                                                  .OnContactPersisted = on_contact_persisted,
                                                  .OnContactRemoved   = on_contact_removed};
-
-// Fixed get_contact_events to be safer with locking
-PyCFunction_DeclareMethodFromModule PhysicsWorld_get_contact_events(PhysicsWorldObject *self,
-                                                                    PyObject *Py_UNUSED(args)) {
-    // --- 1. SNAPSHOT PHASE (Locked) ---
-    SHADOW_LOCK(&self->shadow_lock);
-
-    // Guard: Ensure we aren't reading while Jolt is mid-step updating the buffer
-    BLOCK_UNTIL_NOT_STEPPING(self);
-
-    // Load atomic index (Acquire ensures we see all Listener stores)
-    size_t count = atomic_load_explicit(&self->contact_atomic_idx, memory_order_acquire);
-
-    if (count == 0) {
-        SHADOW_UNLOCK(&self->shadow_lock);
-        return PyList_New(0);
-    }
-
-    if (count > self->contact_max_capacity) {
-        count = self->contact_max_capacity;
-    }
-
-    // Fast copy into local memory so we can drop the lock immediately
-    ContactEvent *scratch = CULV_RAW_MALLOC(count * sizeof(ContactEvent));
-    if (!scratch) {
-        SHADOW_UNLOCK(&self->shadow_lock);
-        return PyErr_NoMemory();
-    }
-
-    memcpy(scratch, self->contact_buffer, count * sizeof(ContactEvent));
-
-    // Reset the index for the next frame
-    atomic_store_explicit(&self->contact_atomic_idx, 0, memory_order_relaxed);
-
-    SHADOW_UNLOCK(&self->shadow_lock);
-
-    // --- 2. BUILD PHASE (Unlocked & FastBuild Integrated) ---
-    PyObject *list = PyList_New((Py_ssize_t)count);
-    if (!list) {
-        CULV_RAW_FREE(scratch);
-        return nullptr;
-    }
-
-    for (size_t i = 0; i < count; i++) {
-        // TSan Fix: Explicitly load the handles from the atomic members in the struct.
-        // We use relaxed because this 'scratch' copy is thread-local and synchronized.
-        uint64_t b1_raw = atomic_load_explicit(&scratch[i].body1, memory_order_relaxed);
-        uint64_t b2_raw = atomic_load_explicit(&scratch[i].body2, memory_order_relaxed);
-
-        /**
-         * OPTIMIZATION: FastBuild_Tuple
-         * 1. fb_from_u64 converts b1_raw and b2_raw
-         * 2. fb_from_float converts impulse and sliding_speed_sq
-         * 3. fb_pack_tuple performs a single O(1) allocation
-         */
-        PyObject *item =
-            FastBuild_Tuple(b1_raw, b2_raw, scratch[i].impulse, scratch[i].sliding_speed_sq);
-
-        if (UNLIKELY(!item)) {
-            Py_DECREF(list);
-            CULV_RAW_FREE(scratch);
-            return nullptr;
-        }
-
-        // PyList_SET_ITEM steals the reference from FastBuild_Tuple
-        PyList_SET_ITEM(list, (Py_ssize_t)i, item);
-    }
-
-    CULV_RAW_FREE(scratch);
-    return list;
-}
-
-PyCFunction_DeclareMethodFromModule PhysicsWorld_get_contact_events_ex(PhysicsWorldObject *self,
-                                                                       PyObject *Py_UNUSED(args)) {
-    // --- 1. SNAPSHOT PHASE ---
-    SHADOW_LOCK(&self->shadow_lock);
-    BLOCK_UNTIL_NOT_STEPPING(self);
-
-    size_t count = atomic_load_explicit(&self->contact_atomic_idx, memory_order_acquire);
-    if (count == 0) {
-        SHADOW_UNLOCK(&self->shadow_lock);
-        return PyList_New(0);
-    }
-
-    if (count > self->contact_max_capacity) {
-        count = self->contact_max_capacity;
-    }
-
-    ContactEvent *scratch = CULV_RAW_MALLOC(count * sizeof(ContactEvent));
-    if (!scratch) {
-        SHADOW_UNLOCK(&self->shadow_lock);
-        return PyErr_NoMemory();
-    }
-
-    memcpy(scratch, self->contact_buffer, count * sizeof(ContactEvent));
-    atomic_store_explicit(&self->contact_atomic_idx, 0, memory_order_relaxed);
-    SHADOW_UNLOCK(&self->shadow_lock);
-
-    // --- 2. KEY INTERNING (Persistent) ---
-    static PyObject *k_bodies = nullptr;
-    static PyObject *k_pos    = nullptr;
-    static PyObject *k_norm   = nullptr;
-    static PyObject *k_str    = nullptr;
-    static PyObject *k_slide  = nullptr;
-    static PyObject *k_mat    = nullptr;
-    static PyObject *k_type   = nullptr;
-
-    if (UNLIKELY(!k_bodies)) {
-        k_bodies = PyUnicode_InternFromString("bodies");
-        k_pos    = PyUnicode_InternFromString("position");
-        k_norm   = PyUnicode_InternFromString("normal");
-        k_str    = PyUnicode_InternFromString("impulse");
-        k_slide  = PyUnicode_InternFromString("slide_sq");
-        k_mat    = PyUnicode_InternFromString("materials");
-        k_type   = PyUnicode_InternFromString("type");
-    }
-
-    // --- 3. BUILD PHASE (FastBuild Engine) ---
-    PyObject *list = PyList_New((Py_ssize_t)count);
-    if (!list) {
-        CULV_RAW_FREE(scratch);
-        return nullptr;
-    }
-
-    for (size_t i = 0; i < count; i++) {
-        ContactEvent *e = &scratch[i];
-
-        // TSan Fix: Explicit relaxed loads for atomic handles
-        uint64_t b1 = atomic_load_explicit(&e->body1, memory_order_relaxed);
-        uint64_t b2 = atomic_load_explicit(&e->body2, memory_order_relaxed);
-
-        /**
-         * OPTIMIZATION: FastBuild_Dict
-         * We compose the nested tuples (pos, normal, bodies, mats)
-         * and the dictionary in a single, readable expression.
-         */
-        PyObject *dict = FastBuild_Dict(
-            k_bodies, FastBuild_Tuple(b1, b2), k_pos, FastBuild_Tuple(e->px, e->py, e->pz), k_norm,
-            FastBuild_Tuple(e->nx, e->ny, e->nz), k_mat, FastBuild_Tuple(e->mat1, e->mat2), k_str,
-            e->impulse, k_slide, e->sliding_speed_sq, k_type, e->type);
-
-        if (UNLIKELY(!dict)) {
-            Py_INCREF(Py_None);
-            PyList_SET_ITEM(list, (Py_ssize_t)i, Py_None);
-            continue;
-        }
-
-        // Steals ref to dict created by FastBuild
-        PyList_SET_ITEM(list, (Py_ssize_t)i, dict);
-    }
-
-    CULV_RAW_FREE(scratch);
-    return list;
-}
-// ContactEvent layout (packed, little-endian):
-// - body1 (uint64)
-// - body2 (uint64)
-// - px, py, pz (float32)
-// - nx, ny, nz (float32)
-// - impulse (float32)
-// - sliding_speed_sq (float32)
-// - mat1 (uint32)
-// - mat2 (uint32)
-// - type (uint32)
-// - _pad (uint32)
-PyCFunction_DeclareMethodFromModule PhysicsWorld_get_contact_events_raw(PhysicsWorldObject *self,
-                                                                        PyObject *Py_UNUSED(args)) {
-    // 1. Phase Guard
-    SHADOW_LOCK(&self->shadow_lock);
-    BLOCK_UNTIL_NOT_STEPPING(self);
-
-    // 2. Atomic Acquire (Publication Visibility)
-    size_t count = atomic_load_explicit(&self->contact_atomic_idx, memory_order_acquire);
-
-    if (count == 0) {
-        SHADOW_UNLOCK(&self->shadow_lock);
-        // Return empty view
-        PyObject *empty = PyBytes_FromStringAndSize("", 0);
-        PyObject *view  = PyMemoryView_FromObject(empty);
-        Py_DECREF(empty);
-        return view;
-    }
-
-    if (count > self->contact_max_capacity) {
-        count = self->contact_max_capacity;
-    }
-
-    // 3. Snapshot Data
-    // We copy into a PyBytes object. This is fast (memcpy) and
-    // ensures the data remains valid even after the next step() resets the
-    // buffer.
-    size_t bytes_size = count * sizeof(ContactEvent);
-    PyObject *raw_bytes =
-        PyBytes_FromStringAndSize((char *)self->contact_buffer, (Py_ssize_t)bytes_size);
-
-    // 4. Reset Index for next frame
-    atomic_store_explicit(&self->contact_atomic_idx, 0, memory_order_relaxed);
-
-    SHADOW_UNLOCK(&self->shadow_lock);
-
-    if (!raw_bytes) {
-        return nullptr;
-    }
-
-    // 5. Wrap in MemoryView
-    // This allows the user to use np.frombuffer(events, dtype=...) without extra
-    // copies
-    PyObject *view = PyMemoryView_FromObject(raw_bytes);
-    Py_DECREF(raw_bytes);
-    return view;
-}
