+++ src/culverin/culverin_contact_listener.h
@@ -1,13 +1,4 @@
 #pragma once
-#include "culverin.h"
+#include "joltc.h"
 
 extern const JPH_ContactListener_Procs contact_procs;
-
-PyCFunction_DeclareMethodFromModule PhysicsWorld_get_contact_events(PhysicsWorldObject *self,
-                                                                    PyObject *args);
-
-PyCFunction_DeclareMethodFromModule PhysicsWorld_get_contact_events_ex(PhysicsWorldObject *self,
-                                                                       PyObject *args);
-
-PyCFunction_DeclareMethodFromModule PhysicsWorld_get_contact_events_raw(PhysicsWorldObject *self,
-                                                                        PyObject *args);
